refactor: split CDC list and search results out of SideMenu
